+++ src/db.rs
@@ -2,9 +2,8 @@ pub mod options;
 
 use std::{fmt::Debug, sync::Arc};
 
+use futures_util::TryStreamExt;
 #[cfg(feature = "in-use-encryption-unstable")]
-use bson::doc;
-use bson::RawDocumentBuf;
 use futures_util::stream::TryStreamExt;
 
 use crate::{
@@ -16,9 +15,9 @@ use crate::{
         ChangeStream,
     },
     client::session::TransactionState,
-    cmap::{conn::PinnedConnectionHandle, Command, RawCommandResponse, StreamDescription},
+    cmap::{conn::PinnedConnectionHandle},
     concern::{ReadConcern, WriteConcern},
-    cursor::{Cursor, CursorSpecification},
+    cursor::{Cursor},
     error::{Error, ErrorKind, Result},
     gridfs::{options::GridFsBucketOptions, GridFsBucket},
     operation::{
@@ -27,8 +26,8 @@ use crate::{
         Create,
         DropDatabase,
         ListCollections,
-        Operation,
         RunCommand,
+        RunCursorCommand,
     },
     options::{
         AggregateOptions,
@@ -489,10 +488,7 @@ impl Database {
             .clone()
             .map(SelectionCriteria::ReadPreference);
         let rcc = RunCommand::new(self.name().to_string(), command, selection_criteria, None)?;
-        let rc_command = RunCursorCommand {
-            run_command: rcc,
-            options,
-        };
+        let rc_command = RunCursorCommand::new(rcc, options)?; 
         let client = self.client();
         client.execute_cursor_operation(rc_command).await
     }
@@ -523,10 +519,7 @@ impl Database {
             _ => {}
         }
         let rcc = RunCommand::new(self.name().to_string(), command, selection_criteria, None)?;
-        let rc_command = RunCursorCommand {
-            run_command: rcc,
-            options,
-        };
+        let rc_command = RunCursorCommand::new(rcc, options)?; 
         let client = self.client();
         client.execute_session_cursor_operation(rc_command, session).await
     }
@@ -667,88 +660,5 @@ impl Database {
     /// Creates a new [`GridFsBucket`] in the database with the given options.
     pub fn gridfs_bucket(&self, options: impl Into<Option<GridFsBucketOptions>>) -> GridFsBucket {
         GridFsBucket::new(self.clone(), options.into().unwrap_or_default())
-    }
-}
-
-pub(super) struct RunCursorCommand<'conn> {
-    run_command: RunCommand<'conn>,
-    options: RunCursorCommandOptions,
-}
-
-impl<'conn> Operation for RunCursorCommand<'conn> {
-    type O = CursorSpecification;
-    type Command = RawDocumentBuf;
-
-    const NAME: &'static str = "run_cursor_command";
-
-    fn build(&mut self, _description: &StreamDescription) -> Result<Command<Self::Command>> {
-        self.run_command.build(_description)
-    }
-
-    fn serialize_command(&mut self, cmd: Command<Self::Command>) -> Result<Vec<u8>> {
-        self.run_command.serialize_command(cmd)
-    }
-
-    fn extract_at_cluster_time(
-        &self,
-        _response: &bson::RawDocument,
-    ) -> Result<Option<bson::Timestamp>> {
-        self.run_command.extract_at_cluster_time(_response)
-    }
-
-    fn handle_error(&self, error: Error) -> Result<Self::O> {
-        Err(error)
-    }
-
-    fn selection_criteria(&self) -> Option<&SelectionCriteria> {
-        self.run_command.selection_criteria()
-    }
-
-    fn is_acknowledged(&self) -> bool {
-        self.run_command.is_acknowledged()
-    }
-
-    fn write_concern(&self) -> Option<&WriteConcern> {
-        self.run_command.write_concern()
-    }
-
-    fn supports_read_concern(&self, _description: &StreamDescription) -> bool {
-        self.run_command.supports_read_concern(_description)
-    }
-
-    fn supports_sessions(&self) -> bool {
-        self.run_command.supports_sessions()
-    }
-
-    fn retryability(&self) -> crate::operation::Retryability {
-        self.run_command.retryability()
-    }
-
-    fn update_for_retry(&mut self) {
-        self.run_command.update_for_retry()
-    }
-
-    fn pinned_connection(&self) -> Option<&PinnedConnectionHandle> {
-        self.run_command.pinned_connection()
-    }
-
-    fn name(&self) -> &str {
-        self.run_command.name()
-    }
-
-    fn handle_response(
-        &self,
-        response: RawCommandResponse,
-        description: &StreamDescription,
-    ) -> Result<Self::O> {
-        let doc = Operation::handle_response(&self.run_command, response, description)?;
-        let cursor_info = bson::from_document(doc)?;
-        Ok(CursorSpecification::new(
-            cursor_info,
-            description.server_address.clone(),
-            self.options.batch_size,
-            self.options.max_time,
-            self.options.comment.clone(),
-        ))
     }
 }
